+++ Snowflake.Data.Tests/UnitTests/Authenticator/WorkflowIdentity/WorkflowIdentityAwsAttestationRetrieverTest.cs
@@ -9,6 +9,7 @@ using System.Threading.Tasks;
 using Amazon.Runtime;
 using Moq;
 using Snowflake.Data.Client;
+using Snowflake.Data.Configuration;
 using Snowflake.Data.Core;
 using Snowflake.Data.Core.Authenticator.WorkflowIdentity;
 using Snowflake.Data.Core.Rest;
