+++ Snowflake.Data.Tests/UnitTests/CredentialManager/SFCredentialManagerFileStorageTest.cs
@@ -1,6 +1,7 @@
 using System;
 using System.IO;
 using Moq;
+using Snowflake.Data.Configuration;
 using Snowflake.Data.Core.CredentialManager.Infrastructure;
 using Snowflake.Data.Core.Tools;
 using Snowflake.Data.Tests.Util;
