+++ Snowflake.Data.Tests/UnitTests/Tools/PlatformDetectionTest.cs
@@ -4,6 +4,7 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using Xunit;
 using RichardSzalay.MockHttp;
+using Snowflake.Data.Configuration;
 using Snowflake.Data.Core.Tools;
 using Snowflake.Data.Tests.Util;
