+++ Snowflake.Data/Core/Authenticator/WorkflowIdentity/WorkflowIdentityAzureAttestationRetriever.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Net.Http;
 using System.Web;
+using Snowflake.Data.Configuration;
 using Snowflake.Data.Core.Extensions;
 using Snowflake.Data.Core.Rest;
 using Snowflake.Data.Core.Tools;
