+++ Snowflake.Data/Core/CredentialManager/Infrastructure/SFCredentialManagerFileStorage.cs
@@ -1,5 +1,6 @@
 using System;
 using System.IO;
+using Snowflake.Data.Configuration;
 using Snowflake.Data.Core.Tools;
 using Snowflake.Data.Log;
